VESA mode is working!
Super excited! This took forever. It now works (may only work for this 
device (QEMU) since the screen memory address is hardcoded) but I works 
and I can draw to screen using RGB and it's awesome and now I can make a 
raytracer/marcher and actually get to coding content